requirements.txt
attrs>=23.1
fairscale>=0.4.13
fire >= 0.6
hydra-core>=1.3.0
numpy>=1.26
omegaconf>=2.3
pandas>=2.2
pyjwt>=2.8
pyyaml>=6.0
requests>=2.31
retry==0.9.2
sentencepiece>=0.2.0
scipy>=1.13
tiktoken>=0.6.0
transformers>=4.39
tqdm >= 4.66